[READY] Coverage for c++ code

### Overview

This PR adds support for sending c++ coverage data (via gcov) to codecov.

The basic approach is to just enable building with gcov via the `-coverage` flag, and to tell `codecov` to ignore the `tests` and `BoostParts` directories.

Please excuse the insane number of commits, it was a big old slog to get this to actually work. I will squash when we're happy with it.

I've pulled out some highlights of what's included in this change:

### `build-dir` switch

I merged a change from my fork to add `--build-dir` to `build.py`. While I find this generally useful (as it reduces the rebuild time when developing and updating ycmd, without having to go through the full instructions), but it is also required here so that we keep the gcov data files around for `codecov` to pick them up.

This could really have been a separate PR, so that's still an option if you guys prefer.

### Exclusions

We don't care about `tests` and `BoostParts` as these are third-party libraries and test code itself. We exclude them from codecov using its difficult to find `ignore` yaml entry.

### Windows

As far as I can tell, there is no support for MSVC coverage in codecov. I understand that it is enabled with the `/Profile` flag to the compiler. I don't think we have os-switched code in the c++ layer, so I don't expect this to be an issue.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/valloric/ycmd/636)
<!-- Reviewable:end -->

+++ build.py
@@ -264,12 +264,25 @@ def ParseArguments():
                        action = 'store_true',
                        help   = 'Enable all supported completers',
                        dest   = 'all_completers' )
+  parser.add_argument( '--enable-coverage',
+                       action = 'store_true',
+                       help   = 'For developers: Enable gcov coverage for the '
+                                'c++ module' )
   parser.add_argument( '--enable-debug',
                        action = 'store_true',
                        help   = 'For developers: build ycm_core library with '
                                 'debug symbols' )
+  parser.add_argument( '--build-dir',
+                       help   = 'For developers: perform the build in the '
+                                'specified directory, and do not delete the '
+                                'build output. This is useful for incremental '
+                                'builds, and required for coverage data' )
 
   args = parser.parse_args()
+
+  if args.enable_coverage:
+    # We always want a debug build when running with coverage enabled
+    args.enable_debug = True
 
   if ( args.system_libclang and
        not args.clang_completer and
@@ -292,6 +305,10 @@ def GetCmakeArgs( parsed_args ):
 
   if parsed_args.enable_debug:
     cmake_args.append( '-DCMAKE_BUILD_TYPE=Debug' )
+
+  # coverage is not supported for c++ on MSVC
+  if not OnWindows() and parsed_args.enable_coverage:
+    cmake_args.append( '-DCMAKE_CXX_FLAGS=-coverage' )
 
   use_python2 = 'ON' if PY_MAJOR == 2 else 'OFF'
   cmake_args.append( '-DUSE_PYTHON2=' + use_python2 )
@@ -339,7 +356,17 @@ def ExitIfYcmdLibInUseOnWindows():
 
 
 def BuildYcmdLib( args ):
-  build_dir = mkdtemp( prefix = 'ycm_build_' )
+  if args.build_dir:
+    build_dir = os.path.abspath( args.build_dir )
+
+    if os.path.exists( build_dir ):
+      print( 'The supplied build directory ' + build_dir + ' exists, '
+             'deleting it.' )
+      rmtree( build_dir, ignore_errors = OnTravisOrAppVeyor() )
+
+    os.makedirs( build_dir )
+  else:
+    build_dir = mkdtemp( prefix = 'ycm_build_' )
 
   try:
     full_cmake_args = [ '-G', GetGenerator( args ) ]
@@ -376,7 +403,11 @@ def BuildYcmdLib( args ):
       RunYcmdTests( build_dir )
   finally:
     os.chdir( DIR_OF_THIS_SCRIPT )
-    rmtree( build_dir, ignore_errors = OnTravisOrAppVeyor() )
+
+    if args.build_dir:
+      print( 'The build files are in: ' + build_dir )
+    else:
+      rmtree( build_dir, ignore_errors = OnTravisOrAppVeyor() )
 
 
 def BuildOmniSharp():
